+++ packages/core/tests/validation.test.ts
@@ -245,8 +245,8 @@ describe('validation', () => {
 
   describe('isPlugin', () => {
     const validPlugins = [
-      { configure: jest.fn() },
-      { configure: jest.fn(), additional: jest.fn() }
+      { install: jest.fn(), start: jest.fn() },
+      { install: jest.fn(), start: jest.fn(), additional: jest.fn() }
     ]
 
     it.each(validPlugins)('passes validation with %s', (value) => {