+++ packages/platforms/browser/lib/auto-instrumentation/resource-load-plugin.ts
@@ -1,5 +1,8 @@
-import type { SpanContextStorage, InternalConfiguration, Plugin, SpanFactory } from '@bugsnag/core-performance'
+import type { SpanContextStorage, Plugin, SpanFactory, PluginContext, Logger } from '@bugsnag/core-performance'
 import type { BrowserConfiguration } from '../config'
+import { defaultNetworkRequestCallback } from '@bugsnag/request-tracker-performance'
+import type { NetworkRequestCallback } from '@bugsnag/request-tracker-performance'
+import type { BrowserNetworkRequestInfo } from './network-request-plugin'
 
 interface ResourceTiming extends PerformanceResourceTiming {
   responseStatus?: number // https://developer.mozilla.org/en-US/docs/Web/API/PerformanceResourceTiming/responseStatus
@@ -40,8 +43,23 @@ export class ResourceLoadPlugin implements Plugin<BrowserConfiguration> {
     private readonly PerformanceObserverClass: typeof PerformanceObserver
   ) {}
 
-  configure (configuration: InternalConfiguration<BrowserConfiguration>) {
+  private enabled: boolean = false
+  private logger: Logger = { debug: console.debug, warn: console.warn, info: console.info, error: console.error }
+  private networkRequestCallback: NetworkRequestCallback<BrowserNetworkRequestInfo> = defaultNetworkRequestCallback
+
+  install (context: PluginContext<BrowserConfiguration>) {
     if (!resourceLoadSupported(this.PerformanceObserverClass)) return
+
+    const { logger, networkRequestCallback } = context.configuration
+
+    if (logger) this.logger = logger
+    if (networkRequestCallback) this.networkRequestCallback = networkRequestCallback
+
+    this.enabled = true
+  }
+
+  start () {
+    if (!this.enabled) return
 
     const observer = new this.PerformanceObserverClass((list) => {
       const entries = list.getEntries() as ResourceTiming[]
@@ -54,12 +72,12 @@ export class ResourceLoadPlugin implements Plugin<BrowserConfiguration> {
         const parentContext = this.spanContextStorage.first
 
         if (parentContext) {
-          const networkRequestInfo = configuration.networkRequestCallback({ url: entry.name, type: entry.initiatorType })
+          const networkRequestInfo = this.networkRequestCallback({ url: entry.name, type: entry.initiatorType })
 
           if (!networkRequestInfo) return
 
           if (typeof networkRequestInfo.url !== 'string') {
-            configuration.logger.warn(`expected url to be a string following network request callback, got ${typeof networkRequestInfo.url}`)
+            this.logger.warn(`expected url to be a string following network request callback, got ${typeof networkRequestInfo.url}`)
             return
           }
 
@@ -69,7 +87,7 @@ export class ResourceLoadPlugin implements Plugin<BrowserConfiguration> {
             url.search = ''
             name = url.href
           } catch (err) {
-            configuration.logger.warn(`Unable to parse URL returned from networkRequestCallback: ${networkRequestInfo.url}`)
+            this.logger.warn(`Unable to parse URL returned from networkRequestCallback: ${networkRequestInfo.url}`)
             return
           }
 
@@ -104,7 +122,7 @@ export class ResourceLoadPlugin implements Plugin<BrowserConfiguration> {
     try {
       observer.observe({ type: 'resource', buffered: true })
     } catch (err) {
-      configuration.logger.warn('Unable to get previous resource loads as buffered observer not supported, only showing resource loads from this point on')
+      this.logger.warn('Unable to get previous resource loads as buffered observer not supported, only showing resource loads from this point on')
       observer.observe({ entryTypes: ['resource'] })
     }
   }